+++ lib/sidekiq/history/middleware.rb
@@ -10,36 +10,38 @@ module Sidekiq
       private
 
       def call_with_sidekiq_history(worker, msg, queue)
-        worker_history = new_history(worker)
+        worker_sratus = new_status
 
         yield
       rescue StandardError => e
-        worker_history[:status] = :failure
-        worker_history[:exception] = e.class.to_s
-        worker_history[:error] = e.message
-        worker_history[:backtrace] = e.backtrace
+        worker_sratus[:failed] = 1
+        worker_sratus[:passed] = 0
 
         raise e
       ensure
-        save_entry_for_worker(worker_history, worker)
+        save_entry_for_worker(worker_sratus, worker)
       end
 
-      def new_history(worker)
+      def new_status
         # worker.sidekiq_options_hash
         {
-          status: :success,
-          error: '',
-          exception: '',
-          backtrace: '',
-          name: worker.class.to_s,
-          timestamp: Time.now.iso8601
+          failed: 0,
+          passed: 1
         }
       end
 
-      def save_entry_for_worker(worker_history, worker)
+      def save_entry_for_worker(worker_sratus, worker)
         Sidekiq.redis do |redis|
-          history = 'sidekiq:history'
-          redis.lpush(history, Sidekiq.dump_json(worker_history))
+          history = "sidekiq:history:#{Time.now.utc.to_date}"
+          value = redis.hget(history, worker.class.to_s)
+
+          if value
+            summary = Sidekiq.load_json(value).symbolize_keys
+            worker_sratus[:failed] = worker_sratus[:failed] + summary[:failed]
+            worker_sratus[:passed] = worker_sratus[:passed] + summary[:passed]
+          end
+
+          redis.hset(history, worker.class.to_s, Sidekiq.dump_json(worker_sratus))
         end
       end
     end