+++ test/test_sidekiq/middleware_test.rb
@@ -3,24 +3,25 @@ require 'minitest_helper'
 module Sidekiq
   module History
     describe 'Middleware' do
-      it 'records history for passed workers' do
+      before do
+        Sidekiq.redis { |c| c.flushdb }
+      end
+
+      let(:history) { "sidekiq:history:#{Time.now.utc.to_date}" }
+
+      it 'records history json for passed worker' do
         middlewared {}
 
         entry = Sidekiq.redis do |redis|
-          redis.lrange('sidekiq:history', 0, -1)
+          redis.hgetall(history)
         end
+        actual = Sidekiq.load_json(entry['HistoryWorker']).symbolize_keys
 
-        actual = Sidekiq.load_json(entry[0]).symbolize_keys
-
-        assert_equal 'success', actual[:status]
-        assert_equal 'HistoryWorker', actual[:name]
-
-        assert_empty actual[:error]
-        assert_empty actual[:backtrace]
-        assert_empty actual[:exception]
+        assert_equal 1, actual[:passed]
+        assert_equal 0, actual[:failed]
       end
 
-      it 'records history for failed workers' do
+      it 'records history json for passed worker' do
         begin
           middlewared do
             raise StandardError.new('failed')
@@ -29,16 +30,31 @@ module Sidekiq
         end
 
         entry = Sidekiq.redis do |redis|
-          redis.lrange('sidekiq:history', 0, -1)
+          redis.hgetall(history)
+        end
+        actual = Sidekiq.load_json(entry['HistoryWorker']).symbolize_keys
+
+        assert_equal 0, actual[:passed]
+        assert_equal 1, actual[:failed]
+      end
+
+      it 'records history for any workers' do
+        middlewared {}
+        middlewared {}
+        begin
+          middlewared do
+            raise StandardError.new('failed')
+          end
+        rescue
         end
 
-        actual = Sidekiq.load_json(entry[0]).symbolize_keys
+        entry = Sidekiq.redis do |redis|
+          redis.hgetall(history)
+        end
+        actual = Sidekiq.load_json(entry['HistoryWorker']).symbolize_keys
 
-        assert_equal 'failure', actual[:status]
-        assert_equal 'HistoryWorker', actual[:name]
-
-        assert_equal 'StandardError', actual[:exception]
-        assert_equal 'failed', actual[:error]
+        assert_equal 2, actual[:passed]
+        assert_equal 1, actual[:failed]
       end
     end
   end