+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* 08.03.2015: Save in redis json with failed and passed jobs for each worker.
+
+  *Anton Davydov*
 * 04.03.2015: Created simple midelware and static page.
 
   *Anton Davydov*